+++ db/src/main/java/com/sonicbase/server/SnapshotManager.java
@@ -32,11 +32,11 @@ public class SnapshotManager {
   public Logger logger;
 
   public static final int SNAPSHOT_BUCKET_COUNT = 128;
-  public static final long SNAPSHOT_SERIALIZATION_VERSION = 22;
-  public static final long SNAPSHOT_SERIALIZATION_VERSION_22 = 22;
-  public static final long SNAPSHOT_SERIALIZATION_VERSION_21 = 21;
-  public static final long SNAPSHOT_SERIALIZATION_VERSION_20 = 20;
-  public static final long SNAPSHOT_SERIALIZATION_VERSION_19 = 19;
+  public static final short SNAPSHOT_SERIALIZATION_VERSION = 22;
+  public static final short SNAPSHOT_SERIALIZATION_VERSION_22 = 22;
+  public static final short SNAPSHOT_SERIALIZATION_VERSION_21 = 21;
+  public static final short SNAPSHOT_SERIALIZATION_VERSION_20 = 20;
+  public static final short SNAPSHOT_SERIALIZATION_VERSION_19 = 19;
 
   private final DatabaseServer server;
   private long lastSnapshot = -1;
